Package structure fix util -> n5
