+++ src/main/java/org/embl/mobie/io/n5/loaders/N5ImageLoader.java
@@ -27,7 +27,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package org.embl.mobie.io.util.loaders;
+package org.embl.mobie.io.n5.loaders;
 
 import bdv.AbstractViewerSetupImgLoader;
 import bdv.ViewerImgLoader;
