+++ src/main/java/org/embl/mobie/io/n5/loaders/S3ImageLoader.java
@@ -1,4 +1,4 @@
-package org.embl.mobie.io.util.loaders;
+package org.embl.mobie.io.n5.loaders;
 
 public interface S3ImageLoader {
