+++ src/main/java/org/embl/mobie/io/n5/source/LabelSource.java
@@ -1,4 +1,4 @@
-package org.embl.mobie.io.util.source;
+package org.embl.mobie.io.n5.source;
 
 import bdv.util.DefaultInterpolators;
 import bdv.viewer.Interpolation;
