+++ src/main/java/org/embl/mobie/io/ome/zarr/OMEZarrViewer.java
@@ -3,7 +3,7 @@ package org.embl.mobie.io.ome.zarr;
 import bdv.util.BdvFunctions;
 import bdv.util.BdvOptions;
 import bdv.util.BdvStackSource;
-import org.embl.mobie.io.util.source.Sources;
+import org.embl.mobie.io.n5.source.Sources;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.sequence.ViewSetup;
 import net.imglib2.Dimensions;
