+++ src/main/java/org/embl/mobie/io/ome/zarr/util/ZarrArrayCreator.java
@@ -1,6 +1,6 @@
 package org.embl.mobie.io.ome.zarr.util;
 
-import org.embl.mobie.io.util.ArrayCreator;
+import org.embl.mobie.io.n5.util.ArrayCreator;
 import net.imglib2.img.cell.CellGrid;
 import net.imglib2.type.NativeType;
 import org.janelia.saalfeldlab.n5.DataBlock;
