+++ src/main/java/org/embl/mobie/io/ome/zarr/writers/imgplus/WriteSequenceToN5OmeZarr.java
@@ -8,8 +8,8 @@ import bdv.img.cache.SimpleCacheArrayLoader;
 import bdv.img.n5.N5ImageLoader;
 import com.google.gson.GsonBuilder;
 import org.embl.mobie.io.ome.zarr.writers.N5OMEZarrWriter;
-import org.embl.mobie.io.util.DownsampleBlock;
-import org.embl.mobie.io.util.ExportScalePyramid;
+import org.embl.mobie.io.n5.util.DownsampleBlock;
+import org.embl.mobie.io.n5.util.ExportScalePyramid;
 import org.embl.mobie.io.ome.zarr.util.OmeZarrMultiscales;
 import org.embl.mobie.io.ome.zarr.util.ZarrAxes;
 import org.embl.mobie.io.ome.zarr.util.ZarrDatasetAttributes;
