+++ src/main/java/org/embl/mobie/io/openorganelle/OpenOrganelleS3Opener.java
@@ -1,6 +1,6 @@
 package org.embl.mobie.io.openorganelle;
 
-import org.embl.mobie.io.util.openers.S3Opener;
+import org.embl.mobie.io.n5.openers.S3Opener;
 import mpicbg.spim.data.SpimData;
 import net.imglib2.util.Cast;
